feat(storages): added

+++ app/files/models/file.model.go
@@ -6,10 +6,10 @@ import (
 
 type File struct {
 	database.AbstractModel
-	EXT     string `gorm:"type:varchar(10);column:ext;not null"`
-	MIME    string `gorm:"type:varchar(255);column:mime;not null"`
-	Size    uint32 `gorm:"type:bigint;not null"`
-	Storage string `gorm:"type:varchar(255);not null"`
-	Hash    string `gorm:"type:varchar(255);not null"`
+	EXT       string `gorm:"type:varchar(10);column:ext;not null"`
+	MIME      string `gorm:"type:varchar(255);column:mime;not null"`
+	Size      uint64 `gorm:"type:bigint;not null"`
+	Storage   string `gorm:"type:varchar(255);not null"`
+	Hash      string `gorm:"type:varchar(255);not null"`
 	FileUsers *[]FileUser
 }
